+++ src/bittytax/conv/parsers/__init__.py
@@ -41,6 +41,7 @@ from . import (
     eternl,
     etherscan,
     exodus,
+    fatstx,
     ftx,
     gatehub,
     gateio,